+++ Sources/BraintreeVenmo/BTVenmoError.swift
@@ -55,6 +55,9 @@ enum BTVenmoError: Error, CustomNSError, LocalizedError {
     /// 8. Failed to fetch Braintree configuration
     case fetchConfigurationFailed
 
+    /// 9. Enriched Customer Data is disabled
+    case enrichedCustomerDataDisabled
+
     static var errorDomain: String {
         "com.braintreepayments.BTVenmoErrorDomain"
     }
@@ -79,6 +82,8 @@ enum BTVenmoError: Error, CustomNSError, LocalizedError {
             return 7
         case .fetchConfigurationFailed:
             return 8
+        case .enrichedCustomerDataDisabled:
+            return 9
         }
     }
 
@@ -102,6 +107,8 @@ enum BTVenmoError: Error, CustomNSError, LocalizedError {
             return description
         case .fetchConfigurationFailed:
             return "Failed to fetch Braintree configuration."
+        case .enrichedCustomerDataDisabled:
+            return "Cannot collect customer data when ECD is disabled. Enable this feature in the Control Panel to collect this data."
         }
     }
 }