+++ UnitTests/BraintreeVenmoTests/BTConfiguration+Venmo_Tests.swift
@@ -38,4 +38,25 @@ class BTConfiguration_Venmo_Tests: XCTestCase {
 
         XCTAssertEqual(configuration.venmoEnvironment, "rockbox")
     }
+    
+    func testVenmoEnrichedCustomerDataEnabled_returnsEcd() {
+        var configurationJSON = BTJSON(value: [
+            "payWithVenmo": ["enrichedCustomerDataEnabled": true]
+        ])
+        var configuration = BTConfiguration(json: configurationJSON)
+
+        XCTAssertTrue(configuration.isVenmoEnrichedCustomerDataEnabled)
+        
+        configurationJSON = BTJSON(value: [
+            "payWithVenmo": ["enrichedCustomerDataEnabled": false]
+        ])
+        configuration = BTConfiguration(json: configurationJSON)
+
+        XCTAssertFalse(configuration.isVenmoEnrichedCustomerDataEnabled)
+
+        configurationJSON = BTJSON(value: ["payWithVenmo": [:]])
+        configuration = BTConfiguration(json: configurationJSON)
+
+        XCTAssertFalse(configuration.isVenmoEnrichedCustomerDataEnabled)
+    }
 }